fix(techdocs): rename TechDocsHeader to TechDocsPageHeader
